Re-Order Tasks window created
Added an item to the Manage Votes window context menu: "Re-Order Tasks".
This window allows the user to change the order that tasks appear in the
tally output. The default sorting order has been changed from
alphabetical to the order in which tasks appear in the thread, however,
there is a button within the Re-Order Tasks window to alphabetise them.