# Atomic Bomb Comit
So this submits a (mostly) fully functional port of NetTally to Avalonia.

The goal of this port was not to touch any of the existing core, and not to have any dependancies on the Windows flavor of NetTally. This was almost entirely succesfull.

Lots of changes in here, but these are the highlights.

## Converters
Mostly ported over unchanged though they use the Avalonia.IConvert interface instead, and the attributes were removed. Some converters were removed as they were unneeded, including the multi-bool converter and the bool to visibility converter

## Navigation
  Ported over mostly unchanaged, though the Show and ShowDialog methods were converted to work with Avalonia.

## Views
Most of the changes occured here obviously.
### Universal Changes - General
* Nomanclature of Dialog windows changed, removing Window suffix (did this on accident)
* Warning `MessageBox` replaced with new `ShowWarning` dialog using an avalonia compatible MessageBox replacement.

### Universal Changes Xaml
* Xaml is now AXaml
* Avalonia doesn't have `Label` so those are replaced with `TextBlock`
* `x:Name` has been replaced with `Name` (necessary for Avalonia Binding)
* `Visibility` is replaced with `IsVisible` (and often removed)
* `ListBox` bind to `Items` not to `ItemSource`
* Unnecessary coverters have been removed or replaced with avalonia equivalents (mostly booltovisibility, multibooland, invert bool, and notnull converters)
* Binding has been reworked in places to use Avalonia flavors, especially element names binding.
* ToolTips have been moved to their Avalonia equivalent.
*
### Universal Changes CodeBehind
* Avalonia requires a blank constructor, so those have been implemented, they should not be called.
* Avalonia does not automatically build objects in the codebehind based on `x:Name`, so `FindControl<Control>("ControlName")` has been subsituted for those.
* When a event function was reworked it's naming convention was changed to UpperCamalCase like: `ControlName_Event`

### MainWindow
Actually one of the most unchanged on the xaml side.
  * The edit button was removed as that functionality has been moved into the New/Edit Quest Dialog.
  * Hyperlinks were removed
  * Much of the onfocus functionality was removed.
### QuestOptions
Largely rebuilt
* Now contains the functionality for changing the quest name and url
* Now has a 'cancel' button, ideally this would let a user cancel and edit, but for now it only can cancel adding a new quest.
* Reset buttons have been merged into one control.
* Now just takes a reference to the `Quest` in question instead of the entire ViewModel.
* The remaining statics have been moved to a new class and added to the core.
### GlobalOptions
Xaml Rebuilt
* Xaml rebuilt into groupboxes, layout still a little rough.
* Now just takes a reference to the `Option` instead of the entire ViewModel.
* The remaining statics have been moved to a new class and added to the core.

### ManageVotes
Xaml mostly untouched, codebehind almost entirely rewritten.
* Task Management functionality removed for now.
* The Codebehind basically contains a ViewModel for this dialog. This isn't ideal, but I didn't want to change the current project style in this respect.
### ManageTasks
Not ported, removed.

## Config
Ported over unchanged, but not functional. Not sure what the error is at the moment. Did have to add a dependancy to Config.Manager, not sure if that will affect compatability or not.

# Outstanding Issues
This port is functional except for two big outstanding issues. The config/save functionality is not working, and the logging is not working. So far I haven't been able to get to the bottom of that.

There are numerous small layout issues that probably need to be addressed as well, and as always the documentation should probably be improved.